вебурок не срастается с методичкой. попробовал по методичке - не компилируется.

+++ src/main/java/com/example/springlesson3/domain/View/dto/ProductDtoDefault.java
@@ -15,7 +15,7 @@ public class ProductDtoDefault {
 
     private String description;
 
-    private Double cost;
+    private Long cost;
 
     private String imgLink;
 }
